日志大改，先尝试对User和Org的日志记录

+++ src/main/java/scms/Controller/OrgController.java
@@ -8,6 +8,7 @@ import org.springframework.web.bind.annotation.ResponseBody;
 import scms.Dao.DataProcessor;
 import scms.Dao.DatabaseRBTree;
 import scms.Interceptor.BridgeData;
+import scms.Interceptor.WriteLog;
 import scms.Service.DataManager;
 import scms.Service.OnlineManager;
 import scms.Service.UserManager;
@@ -19,6 +20,7 @@ import scms.domain.ServerJson.NoticeData;
 import scms.domain.ServerJson.UserFile;
 
 import java.io.File;
+import java.lang.reflect.WildcardType;
 import java.util.List;
 
 /***
@@ -33,13 +35,21 @@ public class OrgController {
     @RequestMapping("/invite")
     public ReturnJson OrgInvite(@RequestBody GetOrgInviteData data){
         NoticeData notice;
-
         if(data.tips != null){
             notice = NoticeMaker.JoinOrg(data.tips,data.org);
         }else{
              notice = NoticeMaker.JoinOrg(data.org);
         }
-        return UserManager.AddNotice(data.userlist,notice);
+        ReturnJson returnJson = UserManager.AddNotice(data.userlist,notice);
+        UserFile userFile = OnlineManager.GetUserData(BridgeData.getRequestInfo(),0L);
+        WriteLog.writeOrgLog(userFile,returnJson,data.org,"OrgInvite");
+        for (int i = 0; i < data.userlist.size(); i++) {
+            userFile = OnlineManager.GetUserData(data.userlist.get(i), 0L);
+            if (userFile != null) {
+                WriteLog.writeOrgLog(userFile,returnJson,data.org,"被邀请进入组织");
+            }
+        }//被邀请的人也要在日志中记录
+        return returnJson;
     }
     //接收org名称
     // -1.org已经存在
@@ -58,6 +68,7 @@ public class OrgController {
             UserFile user = OnlineManager.GetUserData(BridgeData.getRequestInfo(),1L);//用户添加组织信息
             user.owner.add(file);
             dataProcessor.dataItem.type = user.owner.size()-1;
+            WriteLog.writeOrgLog(user,returnJson,org,"OrgCreate"); //添加日志
         }
         return  returnJson;
     }
@@ -85,6 +96,8 @@ public class OrgController {
                 OnlineManager.GetEventData(org,1L);//给这份文件添加权重
             }
         }
+        UserFile userFile = OnlineManager.GetUserData(BridgeData.getRequestInfo(),0L);
+        WriteLog.writeOrgLog(userFile,res,org,"OrgJoin");
         return res;
     }
 
@@ -93,20 +106,28 @@ public class OrgController {
         //删除某个组织
         ReturnJson res = new ReturnJson(true,"组织删除成功");
         DataProcessor Org = OnlineManager.GetEventData(org,0L);
-        List<Long> userAll = Org.dataItem.users; //得到组织中所有用户的信息
-        for (int i = 0; i < userAll.size(); i++) {
-            UserFile tempUserFile = OnlineManager.GetUserData(userAll.get(i),0L);
-            if(tempUserFile != null) {
-                for (int j = 0; j < tempUserFile.player.size(); j++) {
-                    File temp = tempUserFile.player.get(j);
-                    if (temp.getName().equals(org)) {
-                        temp.delete();
+        List<Long> userAll = null;
+        if(Org == null) {
+            res.res = false;
+            res.state = "组织不存在";
+        }
+        else {
+            userAll = Org.dataItem.users; //得到组织中所有用户的信息
+            for (int i = 0; i < userAll.size(); i++) {
+                UserFile tempUserFile = OnlineManager.GetUserData(userAll.get(i), 0L);
+                if (tempUserFile != null) {
+                    for (int j = 0; j < tempUserFile.player.size(); j++) {
+                        File temp = tempUserFile.player.get(j);
+                        if (temp.getName().equals(org)) {
+                            temp.delete();
+                        }
                     }
+                    WriteLog.writeOrgLog(tempUserFile,res,org,"OrgDelete"); //每个组织内的用户都需要记录日志
                 }
-            }
-        } //删除组织中每个用户的组织文件
-        OnlineManager.RemoveOrg(org);//从在线树中删除组织
-        DatabaseRBTree.Remove(org); //删除组织文件
+            } //删除组织中每个用户的组织文件
+            OnlineManager.RemoveOrg(org);//从在线树中删除组织
+            DatabaseRBTree.Remove(org); //删除组织文件
+        }
         return res;
     }
 }
